Major refactoring
- Changes serviceparser to servicehandler
- Implements Handler ABC for each service
- Reorganizes common code ABC
- Automates parser retrieval and each parser implements parse() for
easy recover_objects
